[feat]: 위경도 데이터 타입 변환 & 강아지 조회 데이터 추가

See merge request s09-bigdata-dist-sub2/S09P22B304!18

+++ backend/src/main/java/io/ssafy/p/j9b304/backend/domain/spot/dto/GetHotSpotResponseDto.java
@@ -11,11 +11,11 @@ import java.math.BigDecimal;
 public class GetHotSpotResponseDto {
     private Long spotId;
     private String name;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
 
     @Builder
-    public GetHotSpotResponseDto(Long spotId, String name, BigDecimal latitude, BigDecimal longitude) {
+    public GetHotSpotResponseDto(Long spotId, String name, double latitude, double longitude) {
         this.spotId = spotId;
         this.name = name;
         this.latitude = latitude;
