+++ backend/src/main/java/io/ssafy/p/j9b304/backend/domain/walk/dto/request/RouteAddRequestDto.java
@@ -15,10 +15,10 @@ public class RouteAddRequestDto {
     Integer sequence;
 
     @Schema(description = "위도")
-    BigDecimal latitude;
+    double latitude;
 
     @Schema(description = "경도")
-    BigDecimal longitude;
+    double longitude;
 
     public Route toEntity(Walk walk, char state) {
         return Route.builder()
